chore: Refactor the leader election (#8130)
* chore: Refactor the leader election


* fix: codegen
